backend/requirements/dev.txt
-r base.txt

black==21.10b0
flake8==4.0.1
django-debug-toolbar==2.2
django-extensions==3.0.5
pylint-django==2.3.0
watchdog==2.1.6
pyyaml==6.0
Werkzeug==2.0.2

pytest==6.2.1
pytest-cov==2.10.1
pytest-django==4.1.0